source_code/requirements.txt
absl-py==1.0.0
advex-uar==0.0.6.dev0
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
autoattack @ git+https://github.com/fra31/auto-attack.git@9b264b52bb65c373373727f532865a5551ab9c02
autograd==1.3
backcall==0.2.0
cachetools==4.2.4
certifi==2022.6.15
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
chardet==3.0.4
charset-normalizer==2.0.9
click @ file:///tmp/build/80754af9/click_1646056590078/work
cloudpickle==2.1.0
cox==0.1.post3
cycler==0.10.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator==4.4.2
defusedxml==0.6.0
dill @ file:///tmp/build/80754af9/dill_1593202311028/work
einops==0.4.1
entrypoints==0.3
filelock==3.7.1
fonttools==4.25.0
frozenlist==1.2.0
fsspec==2021.11.1
future @ file:///home/conda/feedstock_root/build_artifacts/future_1635819506105/work
geotorch==0.3.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
GPUtil==1.4.0
grpcio==1.42.0
huggingface-hub==0.7.0
idna==2.10
imagecorruptions==1.1.2
imageio==2.16.0
importlib-metadata==4.8.2
iopath==0.1.9
ipykernel @ file:///tmp/build/80754af9/ipykernel_1633534655931/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1635944169458/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1601490159889/work
jedi @ file:///tmp/build/80754af9/jedi_1611333729159/work
Jinja2==2.11.3
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter @ file:///tmp/build/80754af9/jupyter_1607700846274/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1598884538475/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1636537028672/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612281846643/work
kornia @ file:///home/conda/feedstock_root/build_artifacts/kornia_1641577184411/work
linformer==0.2.1
llvmlite==0.39.1
lmdb @ file:///home/conda/feedstock_root/build_artifacts/python-lmdb_1644189490449/work
Markdown==3.3.6
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621523467000/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1628003481308/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///tmp/build/80754af9/mistune_1607364877025/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict==5.2.0
munch==2.5.0
munkres==1.1.4
nbconvert==5.5.0
nbformat @ file:///tmp/build/80754af9/nbformat_1602783287752/work
networkx==2.6.3
notebook @ file:///tmp/build/80754af9/notebook_1635411643303/work
numba==0.56.2
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.2
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120906940/work
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
patchify==0.2.3
patsy==0.5.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
portalocker==2.3.2
POT @ file:///home/conda/feedstock_root/build_artifacts/pot_1650605030845/work
pretrainedmodels==0.7.4
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.1
psutil @ file:///tmp/build/80754af9/psutil_1612297992929/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyDeprecate==0.3.1
Pygments @ file:///tmp/build/80754af9/pygments_1600458456400/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636110951836/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytorch-lightning==1.5.4
pytorch-pretrained-vit==0.0.7
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434985866/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1600870028330/work
QtPy==1.9.0
recoloradv==0.0.1
regex==2022.6.2
requests==2.25.1
requests-oauthlib==1.3.0
robustbench @ git+https://github.com/RobustBench/robustbench.git@dca3b0f0d782583c534092430e75d8626ee21dd7
robustness==1.2.1.post2
rsa==4.8
scikit-image==0.19.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370406642/work
scipy @ file:///tmp/build/80754af9/scipy_1618855955483/work
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
statsmodels==0.11.1
submitit==1.4.2
tables==3.7.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1654638215170/work
terminado==0.9.4
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
tifffile==2022.2.9
timm @ file:///home/conda/feedstock_root/build_artifacts/timm_1607370272046/work
tokenizers==0.12.1
torch==1.10.0
torchattacks==3.2.7
torchaudio==0.10.0
torchdiffeq==0.2.3
torchdrift @ git+https://github.com/torchdrift/torchdrift/@bda529a7a26c1a9b14d8d878a75122ece22b95d6
torchmetrics==0.6.1
torchsummary==1.5.1
torchvision==0.11.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094708661/work
tqdm==4.56.2
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
transformers==4.19.4
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.25.11
vit-pytorch==0.35.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1607531506226/work
xmltodict==0.13.0
yacs==0.1.6
yarl==1.7.2
zipp @ file:///tmp/build/80754af9/zipp_1603306325174/work
